+++ optimizations/JPEG/Utilities/MathEx.cs
@@ -5,17 +5,17 @@ namespace JPEG.Utilities
 {
     public static class MathEx
     {
-        public static double Sum(int from, int to, Func<int, double> function)
+        public static float Sum(int from, int to, Func<int, float> function)
             => Enumerable.Range(from, to - from).Sum(function);
 
-        public static double SumByTwoVariables(int from1, int to1, int from2, int to2, Func<int, int, double> function)
+        public static float SumByTwoVariables(int from1, int to1, int from2, int to2, Func<int, int, float> function)
             => Sum(from1, to1, x => Sum(from2, to2, y => function(x, y)));
 
-        public static double LoopByTwoVariables(int from1, int to1, int from2, int to2, Action<int, int> function)
+        public static float LoopByTwoVariables(int from1, int to1, int from2, int to2, Action<int, int> function)
             => Sum(from1, to1, x => Sum(from2, to2, y =>
             {
                 function(x, y);
-                return 0;
+                return 0f;
             }));
     }
 }
